Fix warning on unsigned to signed comparison

+++ planner.cpp
@@ -690,7 +690,7 @@ block->steps_y = labs((target[X_AXIS]-position[X_AXIS]) - (target[Y_AXIS]-positi
   #endif
   delta_mm[Z_AXIS] = (target[Z_AXIS]-position[Z_AXIS])/axis_steps_per_unit[Z_AXIS];
   delta_mm[E_AXIS] = ((target[E_AXIS]-position[E_AXIS])/axis_steps_per_unit[E_AXIS])*volumetric_multiplier[active_extruder]*extrudemultiply/100.0;
-  if ( block->steps_x <=dropsegments && block->steps_y <=dropsegments && block->steps_z <=dropsegments )
+  if ( block->steps_x <=(int)dropsegments && block->steps_y <=(int)dropsegments && block->steps_z <=(int)dropsegments )
   {
     block->millimeters = fabs(delta_mm[E_AXIS]);
   }
